requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
boto3==1.13.19
botocore==1.16.19
cffi==1.14.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
Django==3.0.3
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-s3direct==1.1.5
django-storages==1.9.1
django-unused-media==0.1.13
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
docutils==0.15.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
jmespath==0.10.0
jsonfield==3.1.0
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
mysqlclient==1.4.6
parsel==1.6.0
Pillow==7.1.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
w3lib==1.22.0
wrapt==1.11.2
zope.interface==5.1.0
